+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- database: Read settings from config
+
 ### Fixed
 
 - webapp: No browser history for singe media navigation